@@ -6,18 +6,18 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^1.0.2
-  flutter_modular: ^3.2.1
+  cupertino_icons: ^1.0.3
+  flutter_modular: ^3.2.2+1
   mobx: ^2.0.1
   flutter_mobx: ^2.0.0
   responsive_framework: ^0.1.0
   http: ^0.13.3
-  graphql_flutter: ^4.0.1
+  graphql_flutter: ^5.0.0-nullsafety.5
   
 dev_dependencies:
   flutter_test:
