@@ -5,7 +5,7 @@ import 'package:poc_example_integration/app/modules/register/register_store.dart
 
 class RegisterPage extends StatefulWidget {
   final String title;
-  const RegisterPage({Key key, this.title = "Register"}) : super(key: key);
+  const RegisterPage({Key? key, this.title = "Register"}) : super(key: key);
 
   @override
   _RegisterPageState createState() => _RegisterPageState();
