@@ -9,20 +9,10 @@ part of 'products_store.dart';
 // ignore_for_file: non_constant_identifier_names, unnecessary_brace_in_string_interps, unnecessary_lambdas, prefer_expression_function_bodies, lines_longer_than_80_chars, avoid_as, avoid_annotating_with_dynamic
 
 mixin _$ProductsStore on ProductsStoreBase, Store {
-  final _$productAtom = Atom(name: 'ProductsStoreBase.product');
+  Computed<bool>? _$isDisabledComputed;
 
   @override
-  Product get product {
-    _$productAtom.reportRead();
-    return super.product;
-  }
-
-  @override
-  set product(Product value) {
-    _$productAtom.reportWrite(value, super.product, () {
-      super.product = value;
-    });
-  }
+  bool get isDisabled => (_$isDisabledComputed ??= Computed<bool>(() => super.isDisabled, name: 'ProductsStoreBase.isDisabled')).value;
 
   final _$loadingAtom = Atom(name: 'ProductsStoreBase.loading');
 
@@ -36,6 +26,21 @@ mixin _$ProductsStore on ProductsStoreBase, Store {
   set loading(bool value) {
     _$loadingAtom.reportWrite(value, super.loading, () {
       super.loading = value;
+    });
+  }
+
+  final _$productsAtom = Atom(name: 'ProductsStoreBase.products');
+
+  @override
+  ObservableList<Product>? get products {
+    _$productsAtom.reportRead();
+    return super.products;
+  }
+
+  @override
+  set products(ObservableList<Product>? value) {
+    _$productsAtom.reportWrite(value, super.products, () {
+      super.products = value;
     });
   }
 
@@ -54,8 +59,7 @@ mixin _$ProductsStore on ProductsStoreBase, Store {
     });
   }
 
-  final _$descriptionControllerAtom =
-      Atom(name: 'ProductsStoreBase.descriptionController');
+  final _$descriptionControllerAtom = Atom(name: 'ProductsStoreBase.descriptionController');
 
   @override
   TextEditingController get descriptionController {
@@ -65,8 +69,7 @@ mixin _$ProductsStore on ProductsStoreBase, Store {
 
   @override
   set descriptionController(TextEditingController value) {
-    _$descriptionControllerAtom.reportWrite(value, super.descriptionController,
-        () {
+    _$descriptionControllerAtom.reportWrite(value, super.descriptionController, () {
       super.descriptionController = value;
     });
   }
@@ -89,17 +92,15 @@ mixin _$ProductsStore on ProductsStoreBase, Store {
   final _$saveProdutAsyncAction = AsyncAction('ProductsStoreBase.saveProdut');
 
   @override
-  Future<Graphql.QueryResult> saveProdut() {
+  Future<void> saveProdut() {
     return _$saveProdutAsyncAction.run(() => super.saveProdut());
   }
 
-  final _$ProductsStoreBaseActionController =
-      ActionController(name: 'ProductsStoreBase');
+  final _$ProductsStoreBaseActionController = ActionController(name: 'ProductsStoreBase');
 
   @override
   dynamic setLoading(bool isLoading) {
-    final _$actionInfo = _$ProductsStoreBaseActionController.startAction(
-        name: 'ProductsStoreBase.setLoading');
+    final _$actionInfo = _$ProductsStoreBaseActionController.startAction(name: 'ProductsStoreBase.setLoading');
     try {
       return super.setLoading(isLoading);
     } finally {
@@ -108,13 +109,24 @@ mixin _$ProductsStore on ProductsStoreBase, Store {
   }
 
   @override
+  dynamic addProduct(Product product) {
+    final _$actionInfo = _$ProductsStoreBaseActionController.startAction(name: 'ProductsStoreBase.addProduct');
+    try {
+      return super.addProduct(product);
+    } finally {
+      _$ProductsStoreBaseActionController.endAction(_$actionInfo);
+    }
+  }
+
+  @override
   String toString() {
     return '''
-product: ${product},
 loading: ${loading},
+products: ${products},
 nameController: ${nameController},
 descriptionController: ${descriptionController},
-priceController: ${priceController}
+priceController: ${priceController},
+isDisabled: ${isDisabled}
     ''';
   }
 }
